requirements.txt
langchain
langchain-community
langchain_core
langchain-groq
streamlit
python-dotenv
